+++ packages/terra-framework-docs/src/terra-dev-site/test/embedded-content-consumer/consumers/BasicConsumer.test.jsx
@@ -13,9 +13,6 @@ const BasicConsumer = () => (
     <h2>Embedded Content</h2>
     <p>The following is an embedded content within an iframe.</p>
     <EmbeddedContentConsumer
-      // TODO: Fix site doc so that relative paths can be used everywhere.
-      // For now, use `../../../../` prefix if there is any changes to the providers file so we can see changes during the PR-preview build.
-      // Otherwise, keep the existing `/terra-framework/` prefix so the page will load when viewing the live site.
       src="../../../../#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/basic-provider"
       className={cx('iframe')}
       options={{ iframeAttrs: { id: 'basic-consumer-frame' }, resizeConfig: { scrolling: true, fixedWidth: '100%', fixedHeight: '200px' } }}